+++ modules/replay/util/file_operations.h
@@ -71,8 +71,8 @@ namespace caspar { namespace replay {
 	struct mjpeg_file_header {
 		char							magick[4]; // = 'OMAV'
 		uint8_t							version; // = 1 for version 1, or 2 for version 2
-		size_t							width;
-		size_t							height;
+		uint32_t							width;
+		uint32_t							height;
 		double							fps;
 		caspar::core::field_mode::type	field_mode;
 		boost::posix_time::ptime		begin_timecode;
@@ -105,7 +105,7 @@ namespace caspar { namespace replay {
 	void safe_fclose(mjpeg_file_handle file_handle);
 	void write_index_header(mjpeg_file_handle outfile_idx, const core::video_format_desc* format_desc, boost::posix_time::ptime start_timecode, int audio_channels);
 	void write_index(mjpeg_file_handle outfile_idx, long long offset);
-	long long write_frame(mjpeg_file_handle outfile, size_t width, size_t height, const uint8_t* image, short quality, mjpeg_process_mode mode, chroma_subsampling subsampling, const int32_t* audio_data, size_t audio_data_length);
+	long long write_frame(mjpeg_file_handle outfile, uint32_t width, uint32_t height, const uint8_t* image, short quality, mjpeg_process_mode mode, chroma_subsampling subsampling, const int32_t* audio_data, uint32_t audio_data_length);
 	long long read_index(mjpeg_file_handle infile_idx);
 	long long tell_index(mjpeg_file_handle infile_idx);
 	long long length_index(mjpeg_file_handle infile_idx);
@@ -113,6 +113,6 @@ namespace caspar { namespace replay {
 	long long tell_frame(mjpeg_file_handle infile);
 	int read_index_header(mjpeg_file_handle infile_idx, mjpeg_file_header** header);
 	int read_index_header_ex(mjpeg_file_handle infile_idx, mjpeg_file_header_ex** header);
-	size_t read_frame(mjpeg_file_handle infile, size_t* width, size_t* height, uint8_t** image, size_t* audioSize, int32_t** audio);
+	uint32_t read_frame(mjpeg_file_handle infile, uint32_t* width, uint32_t* height, uint8_t** image, uint32_t* audioSize, int32_t** audio);
 	int seek_frame(mjpeg_file_handle infile, long long offset, uint32_t origin);
 }}
